Added changes in navbar and editor data format
