+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@editorjs/editorjs": "^2.29.1",
     "@editorjs/header": "^2.8.1",
+    "@editorjs/image": "^2.9.0",
     "@editorjs/list": "^1.9.0",
     "@editorjs/paragraph": "^2.11.4",
     "@editorjs/table": "^2.3.0",
