--- src/components/layout/navbar/UserProfile.tsx
@@ -1,11 +1,0 @@
-import Icon from '@/components/icon';
-
-const UserProfile = () => {
-  return (
-    <div className='flex h-10 w-10 items-center justify-center rounded-full'>
-      <Icon name='RiUserFill' />
-    </div>
-  );
-};
-
-export default UserProfile;
