+++ src/components/layout/navbar/notifications/NotificationTabs.tsx
@@ -12,6 +12,7 @@ const NotificationTabs = () => {
         onClick={() => setTab('all')}
       >
         <p className='px-2 font-josefin_Sans text-base'>All</p>
+
         {tab === 'all' && (
           <span className='w-full border-b-2 border-primary-monkeyOrange'></span>
         )}
@@ -22,6 +23,7 @@ const NotificationTabs = () => {
         onClick={() => setTab('comments')}
       >
         <p className='px-2 font-josefin_Sans text-base'>Comments</p>
+
         {tab === 'comments' && (
           <span className='w-full border-b-2 border-primary-monkeyOrange'></span>
         )}
@@ -32,6 +34,7 @@ const NotificationTabs = () => {
         onClick={() => setTab('activity')}
       >
         <p className='px-2 font-josefin_Sans text-base'>Post Activity</p>
+
         {tab === 'activity' && (
           <span className='w-full border-b-2 border-primary-monkeyOrange'></span>
         )}
